New command line argument parsing code
New module `Debian.Debianize.Optparse' is reimplementation
of subset of options, currently parsed via GetOpt in
`Debian.Debianize.Options'.

At this point of time, `Optparse` module is not used, but
correctly complies. Unfortunately, it is big change and
we will be able to substitute `Options' with `Optparse'
after a lot of work.

Signed-off-by: Dmitry Bogatov <KAction@gnu.org>